+++ vars/entry.groovy
@@ -1,99 +1,34 @@
-def call(Map config=[:]){
-    pipeline {
-        agent any
+def call(Map args = [:]) {
+    /*
+    Order:
+        1. Local Settings
+        2. .jenkins.yaml
+        3. Config
+    */
 
-        options {
-            //timestamps()
-            timeout(time: 30, unit: 'MINUTES')
-            buildDiscarder(logRotator(numToKeepStr: '10'))
-        }
+    loadSettings()
 
-        parameters {
-            gitParameter (
-                branch: '', 
-                branchFilter: 'origin/(.*)', 
-                defaultValue: 'master', 
-                listSize: '10', 
-                name: 'GIT_REVISION', 
-                quickFilterEnabled: true, 
-                selectedValue: 'NONE', 
-                sortMode: 'DESCENDING_SMART', 
-                tagFilter: '*', 
-                type: 'PT_BRANCH_TAG', 
-                description: 'Please select a branch or tag to build',
-                useRepository: config.repo)
+    if(!ACTION) { ACTION = "deploy" }
 
-            choice(
-                name: 'ENVIRONMENT',
-                description: 'Please select Environment',
-                choices: 'DEV\nTEST\nUAT\nPRE\nPRD')
+    Config.data['repo']             = args.containsKey('repo')              ?: null
+    Config.data['revision']         = args.containsKey('revision')          ?: GIT_REVISION
+    Config.data['language']         = args.containsKey('language')          ?: "java"
+    Config.data['action']           = ACTION
+    Config.data['build.user']       = BUILD_USER
+    Config.data['env']              = ENVIRONMENT
+    Config.data['credentials.id']   = args.containsKey('credentials.id')    ?: "DefaultGitSCMCredentialsID"
 
-            choice(
-                name: 'ACTION',
-                description: 'Please select action',
-                choices: 'deploy\nrollback')
-        }
+    Config.data += args 
 
-        stages {
-            stage ("Initial Stages") {
-                steps {
-                    script {
-                        log.i "Acquire config data"
-                        // Fecth data
-                        /*
-                        
-                        fc = new FetchConfig(auto)
-                        
-                        fc.getConfig()
-                        */
-                        stagesController(["repo": config.repo,                       
-                            "lang": "java",
-                            "build.command": "mvn",
-                            "build.options": "-U clean -Dmaven.test.skip=true package dependency:tree"])
-                    }
-                }
-            }
-        }
+    println Config.data 
+    println args 
 
-        post {
-            aborted {
-                script {
-                    log.i "Post Action: aborted"
-                }
-            }
+    log.i "Using workspace: " + FIRST_DIR
 
-            always {
-                script {
-                    log.i "Post Action: always"
-                }
-            }
-
-            changed {
-                script {
-                    log.i "Post Action: changed"
-                }
-            }
-
-            failure {
-                script {
-                    currentBuild.result = "FAILURE"
-                    log.i "Post Action: failure"  
-
-                    // Send mail
-                }
-            }
-
-            success {
-                script {
-                    log.i "Post Action: success"
-                }
-            }
-
-            unstable {
-                script {
-                    log.i "Post Action: unstable"
-                }
-            }  
-        }
-    }
+    stagePreProcess()
+    stageGitClone()
+    stageBuild()
+    stageTest()
+    stageDocker()
+    stageKubernetes()
 }
